Create student class
