--- teacher.rb
@@ -1,23 +1,0 @@
-class Person
-    attr_reader :id
-    attr_accessor :name, :age
-
-    def initialize(age, name = 'unknown', parent_permission = true)
-        @id = Random.rand(1...1000)
-        @name = name
-        @age = age
-        @parent_permission = parent_permission
-    end
-
-    private
-
-    def of_age?
-        @age >= 18
-    end   
-    
-    public 
-
-    def can_use_services?
-        of_age? || @parent_permission
-    end
-end
